requirements/base.txt
Django
django-summernote
django-redis
lxml
Pillow
easy-thumbnails
psycopg2
django-solo
dj-pagination
django-admin-sortable2